refactor types to new file
