+++ src/services/scheduler-service.utils.ts
@@ -1,42 +1,9 @@
 import { StatusCodes } from 'http-status-codes';
 
+import { SchedulerJob } from 'src/types/services/scheduler-service';
 import logger from 'utils/logger';
 
 import { HttpError } from '../types/errors';
-import { BaseResponseHttpMetaData } from '../types/services/api-service';
-
-export type SchedulerJob = {
-  name: string;
-  schedule: string;
-  targetUrl: string;
-  retryConfig?: {
-    maxRetries: number;
-    minBackoffDuration: number;
-  };
-  timeout?: number;
-};
-
-export type ListJobsResponse = {
-  jobs: SchedulerJob[];
-} & BaseResponseHttpMetaData;
-
-export type CreateJobRequest = {
-  name: string;
-  description?: string;
-  schedule: string;
-  targetUrl: string;
-  retryConfig?: {
-    maxRetries?: number;
-    minBackoffDuration?: number;
-  };
-  timeout?: number;
-};
-
-export type UpdateJobRequest = Partial<Omit<CreateJobRequest, 'name'>>;
-
-export type CreateJobResponse = {
-  job: SchedulerJob;
-} & BaseResponseHttpMetaData;
 
 export const printJobs = (jobs: SchedulerJob[]): void => {
   if (jobs.length === 0) {
